requirements.txt
absl-py==1.4.0
accelerate==0.18.0
aenum==3.1.15
aiohttp==3.8.4
aiosignal==1.3.1
aniso8601==9.0.1
ansi2html==1.8.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
arrow==1.2.3
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blessed==1.20.0
cachetools==5.3.0
captum==0.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
contourpy==1.0.7
croniter==1.3.14
cycler==0.11.0
Cython==0.29.34
datasets==2.10.1
dateutils==0.6.12
deepdiff==6.3.0
dill==0.3.6
dm-tree==0.1.8
docker-pycreds==0.4.0
einops==0.6.1
enum-compat==0.0.3
fastapi==0.88.0
filelock==3.12.0
Flask==2.2.3
Flask-RESTful==0.3.9
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2025.5.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
h11==0.14.0
h5py==3.10.0
hf-xet==1.1.5
huggingface-hub==0.33.2
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
lightly==1.4.10
lightly-utils==0.0.2
lightning==2.0.1.post0
lightning-bolts==0.6.0.post1
lightning-cloud==0.5.34
lightning-utilities==0.8.0
lit==16.0.1
llvmlite==0.41.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
numba==0.58.1
numpy==1.24.3
nvgpu==0.10.0
nvidia-dali-cuda110==1.30.0
oauthlib==3.2.2
omegaconf==2.3.0
ordered-set==4.1.0
packaging==23.1
pandas==2.0.0
pathtools==0.1.2
Pillow==9.3.0
protobuf==3.20.3
psutil==5.9.5
ptflops==0.7.4
pyarrow==20.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
Pygments==2.15.1
PyJWT==2.6.0
pynndescent==0.5.10
pynvml==11.4.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.2
pytz==2023.3
PyYAML==6.0
readchar==4.0.5
regex==2024.11.6
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.3.4
rsa==4.9
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
sentry-sdk==1.32.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
solo-learn @ file:///home/model-server/solo-learn
soupsieve==2.4.1
starlette==0.22.0
starsessions==1.3.0
sympy==1.11.1
tabulate==0.9.0
tbb==2021.10.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
termcolor==2.2.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
timm==0.9.8
tokenizers==0.13.3
torch==2.0.0+cu117
torch-model-archiver==0.7.1
torchdata==0.6.0
torchmetrics==0.11.4
torchserve==0.7.1
torchtext==0.15.1
torchvision==0.15.1+cu117
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
umap-learn==0.5.4
urllib3==1.26.15
uvicorn==0.21.1
wandb==0.15.12
wcwidth==0.2.6
websocket-client==1.5.1
websockets==11.0.2
Werkzeug==2.2.3
xxhash==3.5.0
yarl==1.9.1
zipp==3.15.0
